:construction: GI-4 Cast to float instead of int

+++ src/models/api_responses/dodo/productivity_balance.py
@@ -8,5 +8,5 @@ __all__ = ('UnitProductivityBalanceStatisticsReport',)
 class UnitProductivityBalanceStatisticsReport(BaseModel):
     unit_uuid: UUID
     sales_per_labor_hour: int
-    orders_per_labor_hour: int
+    orders_per_labor_hour: float
     stop_sale_duration_in_seconds: int
